Moved Ticket tests into new structure
