+++ tests/protocols/gmpv208/entities/tickets/test_get_ticket.py
@@ -19,26 +19,23 @@
 from gvm.errors import RequiredArgument
 
 
-class GmpCloneTicketTestCase:
-    def test_clone(self):
-        self.gmp.clone_ticket('t1')
+class GmpGetTicketTestMixin:
+    def test_get_ticket(self):
+        self.gmp.get_ticket('t1')
 
         self.connection.send.has_been_called_with(
-            '<create_ticket>' '<copy>t1</copy>' '</create_ticket>'
+            '<get_tickets ticket_id="t1"/>'
         )
 
-        self.gmp.clone_ticket(ticket_id='t1')
+        self.gmp.get_ticket(ticket_id='t1')
 
         self.connection.send.has_been_called_with(
-            '<create_ticket>' '<copy>t1</copy>' '</create_ticket>'
+            '<get_tickets ticket_id="t1"/>'
         )
 
-    def test_missing_id(self):
+    def test_get_ticket_missing_user_id(self):
         with self.assertRaises(RequiredArgument):
-            self.gmp.clone_ticket('')
+            self.gmp.get_ticket(ticket_id=None)
 
         with self.assertRaises(RequiredArgument):
-            self.gmp.clone_ticket(None)
-
-        with self.assertRaises(RequiredArgument):
-            self.gmp.clone_ticket(ticket_id=None)
+            self.gmp.get_ticket('')
