+++ tests/protocols/gmpv208/testcmds/__init__.py
@@ -18,12 +18,6 @@
 
 # pylint: disable=no-member
 
-from .test_modify_ticket import GmpModifyTicketTestCase
-from .test_clone_ticket import GmpCloneTicketTestCase
-from .test_create_ticket import GmpCreateTicketTestCase
-from .test_delete_ticket import GmpDeleteTicketTestCase
-from .test_get_ticket import GmpGetTicketTestCase
-from .test_get_tickets import GmpGetTargetsTestCase
 from .test_get_vulnerabilities import GmpGetVulnerabilitiesTestCase
 from .test_get_vulnerability import GmpGetVulnerabilityTestCase
 from .test_protocol_version import GmpProtocolVersionTestCase
